project/pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.16.4",
    "beautifulsoup4>=4.13.4",
    "crawl4ai>=0.7.2",
    "fastapi>=0.116.1",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "selenium>=4.34.2",
    "sqlalchemy>=2.0.42",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ipython>=8.37.0",
    "notebook>=7.4.4",
    "ruff>=0.12.5",
]
